requirements.txt
torch==2.1.1
numpy
scipy
matplotlib
soundfile
tqdm
flash-attn
dearpygui
sounddevice
scikit-learn
umap-learn
